add github properties
