+++ .gitignore
@@ -23,6 +23,7 @@ build/
 
 # Local configuration file (sdk path, etc)
 local.properties
+github.properties
 
 # Proguard folder generated by Eclipse
 proguard/
