+++ README.md
@@ -2,3 +2,11 @@
 XMTP client SDK for Android applications written in Kotlin.
 
 **Coming soon!** This is a work in progress and should not be used yet. Follow along in the [tracking issue](https://github.com/xmtp/xmtp-android/issues/1) for updates on progress.
+
+### Setup
+
+```bash
+cp github.properties.example github.properties
+```
+
+Fill in with your GitHub credentials. Token must have package read access.
