+++ example/src/androidTest/java/org/xmtp/android/example/MainActivityInstrumentedTest.kt
@@ -1,0 +1,19 @@
+package org.xmtp.android.example
+
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.*
+
+@RunWith(AndroidJUnit4::class)
+class MainActivityInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("org.xmtp.android.example", appContext.packageName)
+    }
+}
