+++ example/src/main/res/values/strings.xml
