+++ github.properties
@@ -1,0 +1,4 @@
+# See more information at https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-gradle-registry
+
+gpr.user="nplasterer"
+gpr.key="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"
