+++ settings.gradle
@@ -10,11 +10,17 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
+        def githubProperties = new Properties()
+        githubProperties.load(new FileInputStream(file("github.properties")))
         maven {
             url = uri("https://maven.pkg.github.com/xmtp/proto")
+            credentials {
+                username = githubProperties.getProperty("gpr.user")
+                password = githubProperties.getProperty("gpr.key")
+            }
         }
     }
 }
-rootProject.name = "XMTP Android"
-include ':app'
-include ':message'
+rootProject.name = "xmtp-android"
+include ':example'
+include ':library'
